Renamed keyboard folder to timer
